refactor(server): #125 — extract file-transfer handlers into routes/files.rs
